+++ zarrs_storage/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  - Add `ByteRange::new` and `From` for `RangeBounds<u64>`
- - Add `PerformanceMetricsStorageAdapter::reset()`
+ - Add `PerformanceMetricsStorageAdapter::{keys_erased,reset}()`
 
 ### Changed
  - Bump `unsafe_cell_slice` to 0.2.0